formulagan/requirements.txt
click==7.1.2
Distance==0.1.3
keras==2.6.0
keras_visualizer==2.4
matplotlib==3.4.3
nltk==3.6.5
numpy
opencv_python_headless==4.5.5.64
pandas==1.3.4
Pillow==9.2.0
pix2tex==0.0.26
scipy==1.7.1
six==1.15.0
streamlit==1.8.1
sympy==1.9
tensorflow==2.6.0
tqdm==4.62.3
visualkeras==0.0.2